数据库配置样例完成，后续公共按照此模板展开.
