+++ .idea/artifacts/hzau_bbs_web_war.xml
@@ -1,0 +1,8 @@
+<component name="ArtifactManager">
+  <artifact type="war" name="hzau-bbs-web:war">
+    <output-path>$PROJECT_DIR$/hzau-bbs-web/target</output-path>
+    <root id="archive" name="hzau-bbs-web.war">
+      <element id="artifact" artifact-name="hzau-bbs-web:war exploded" />
+    </root>
+  </artifact>
+</component>
