+++ .idea/encodings.xml
@@ -2,6 +2,7 @@
 <project version="4">
   <component name="Encoding">
     <file url="file://$PROJECT_DIR$" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/hzau-bbs-core" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/hzau-bbs-data" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/hzau-bbs-service" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/hzau-bbs-web" charset="UTF-8" />
