--- .idea/libraries/Maven__javax_persistence_persistence_api_1_0_2.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: javax.persistence:persistence-api:1.0.2">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/javax/persistence/persistence-api/1.0.2/persistence-api-1.0.2.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/javax/persistence/persistence-api/1.0.2/persistence-api-1.0.2-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/javax/persistence/persistence-api/1.0.2/persistence-api-1.0.2-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
