--- .idea/libraries/Maven__mysql_mysql_connector_java_5_1_46.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: mysql:mysql-connector-java:5.1.46">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/mysql/mysql-connector-java/5.1.46/mysql-connector-java-5.1.46.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/mysql/mysql-connector-java/5.1.46/mysql-connector-java-5.1.46-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/mysql/mysql-connector-java/5.1.46/mysql-connector-java-5.1.46-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
