--- .idea/libraries/Maven__net_bytebuddy_byte_buddy_1_7_9.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: net.bytebuddy:byte-buddy:1.7.9">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/net/bytebuddy/byte-buddy/1.7.9/byte-buddy-1.7.9.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/net/bytebuddy/byte-buddy/1.7.9/byte-buddy-1.7.9-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/net/bytebuddy/byte-buddy/1.7.9/byte-buddy-1.7.9-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
