--- .idea/libraries/Maven__org_apache_tomcat_tomcat_jdbc_8_5_27.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: org.apache.tomcat:tomcat-jdbc:8.5.27">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/apache/tomcat/tomcat-jdbc/8.5.27/tomcat-jdbc-8.5.27.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/org/apache/tomcat/tomcat-jdbc/8.5.27/tomcat-jdbc-8.5.27-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/apache/tomcat/tomcat-jdbc/8.5.27/tomcat-jdbc-8.5.27-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
