--- .idea/libraries/Maven__org_aspectj_aspectjrt_1_8_13.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: org.aspectj:aspectjrt:1.8.13">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/aspectj/aspectjrt/1.8.13/aspectjrt-1.8.13.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/org/aspectj/aspectjrt/1.8.13/aspectjrt-1.8.13-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/aspectj/aspectjrt/1.8.13/aspectjrt-1.8.13-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
