--- .idea/libraries/Maven__org_hibernate_validator_hibernate_validator_6_0_11_Final.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: org.hibernate.validator:hibernate-validator:6.0.11.Final">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/hibernate/validator/hibernate-validator/6.0.11.Final/hibernate-validator-6.0.11.Final.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/org/hibernate/validator/hibernate-validator/6.0.11.Final/hibernate-validator-6.0.11.Final-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/hibernate/validator/hibernate-validator/6.0.11.Final/hibernate-validator-6.0.11.Final-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
