--- .idea/libraries/Maven__org_springframework_boot_spring_boot_2_1_0_M2.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: org.springframework.boot:spring-boot:2.1.0.M2">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/springframework/boot/spring-boot/2.1.0.M2/spring-boot-2.1.0.M2.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/org/springframework/boot/spring-boot/2.1.0.M2/spring-boot-2.1.0.M2-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/springframework/boot/spring-boot/2.1.0.M2/spring-boot-2.1.0.M2-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
