--- .idea/libraries/Maven__org_springframework_spring_tx_4_3_14_RELEASE.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: org.springframework:spring-tx:4.3.14.RELEASE">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/springframework/spring-tx/4.3.14.RELEASE/spring-tx-4.3.14.RELEASE.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/org/springframework/spring-tx/4.3.14.RELEASE/spring-tx-4.3.14.RELEASE-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/springframework/spring-tx/4.3.14.RELEASE/spring-tx-4.3.14.RELEASE-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
