+++ hzau-bbs-core/src/main/resources/application.properties
@@ -1,0 +1,22 @@
+# datasource
+datasource.hzau.master.url=jdbc:mysql://localhost:3306/hzau_bbs?useUnicode=true&characterEncoding=utf8&zeroDateTimeBehavior=CONVERT_TO_NULL&allowMultiQueries=true
+datasource.hzau.master.username=root
+datasource.hzau.master.password=root
+datasource.hzau.master.type=com.alibaba.druid.pool.DruidDataSource
+datasource.hzau.master.driver=com.mysql.cj.jdbc.Driver
+datasource.hzau.master.initial-size=5
+datasource.hzau.master.max-active=100
+#  filters: stat
+#  maxActive: 90
+#  initialSize: 10
+#  maxWait: 60000
+#  minIdle: 1
+#  maxIdle: 90
+#  timeBetweenEvictionRunsMillis: 60000
+#  minEvictableIdleTimeMillis: 300000
+#  validationQuery: select 'x'
+#  testWhileIdle: true
+#  testOnBorrow: false
+#  testOnReturn: false
+#  poolPreparedStatements: true
+#  maxOpenPreparedStatements: 20
