+++ hzau-bbs-data/hzau-bbs.sql
@@ -1,10 +1,22 @@
 USE hzau_bbs;
 
+DROP TABLE user;
+
 CREATE TABLE user (
-id INTEGER PRIMARY KEY,
-username VARCHAR (20) NOT NULL,
-number VARCHAR (15) NOT NULL,
-email VARCHAR (50) NOT NULL,
-update_time TIMESTAMP DEFAULT now()
-create_time TIMESTAMP NOT NULL
-);
+id BIGINT(20) PRIMARY KEY AUTO_INCREMENT
+COMMENT '主键',
+username VARCHAR (20) NOT NULL DEFAULT ''
+COMMENT '用户名',
+number VARCHAR (15) NOT NULL UNIQUE
+COMMENT '学号',
+email VARCHAR (50) NOT NULL DEFAULT ''
+COMMENT '邮箱',
+mobile VARCHAR (15) NOT NULL DEFAULT ''
+COMMENT '手机号码',
+enable BOOLEAN NOT NULL DEFAULT FALSE
+COMMENT '是否有效',
+update_time TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP  ON UPDATE CURRENT_TIMESTAMP
+COMMENT '更新时间',
+create_time TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP
+COMMENT '创建时间'
+) ENGINE=InnoDB DEFAULT CHARSET=utf8;
