--- hzau-bbs-data/src/main/resources/application.yml
@@ -1,26 +1,0 @@
-# datasource
-datasource:
-  hzau:
-    master:
-      url:
-      username:
-      password:
-      type: com.alibaba.druid.pool.DruidDataSource
-      driver: com.mysql.cj.jdbc.Driver
-      initial-size: 5
-      max-active: 100
-
-#  filters: stat
-#  maxActive: 90
-#  initialSize: 10
-#  maxWait: 60000
-#  minIdle: 1
-#  maxIdle: 90
-#  timeBetweenEvictionRunsMillis: 60000
-#  minEvictableIdleTimeMillis: 300000
-#  validationQuery: select 'x'
-#  testWhileIdle: true
-#  testOnBorrow: false
-#  testOnReturn: false
-#  poolPreparedStatements: true
-#  maxOpenPreparedStatements: 20
