+++ hzau-bbs-data/src/main/resources/org/feidian/hzau/bbs/data/dao/mapper/UserMapper.xml
@@ -1,0 +1,35 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper
+        PUBLIC "-//ibatis.apache.org//DTD Config 3.1//EN"
+        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="org.feidian.hzau.bbs.data.dao.mapper.UserMapper">
+    <insert id="insert">
+        INSERT INTO user (
+            username,
+            number,
+            email,
+            enable,
+            update_time,
+            create_time
+        ) VALUES (
+            #{user.username},
+            #{user.number},
+            #{user.email},
+            1,
+            #{user.updateTime},
+            #{user.createTime}
+        )
+    </insert>
+
+    <select id="listByIds" resultType="org.feidian.hzau.bbs.data.entity.User">
+        SELECT *
+        FROM user
+        WHERE enable = #{enable} AND id IN
+        <foreach collection="ids" item="id" open="(" separator="," close=")">
+            #{id}
+        </foreach>
+    </select>
+    <select id="listAll" resultType="org.feidian.hzau.bbs.data.entity.User">
+        SELECT * FROM user
+    </select>
+</mapper>
