+++ hzau-bbs-service/pom.xml
@@ -10,6 +10,10 @@
     <modelVersion>4.0.0</modelVersion>
 
     <artifactId>hzau-bbs-service</artifactId>
+    <version>1.0-SNAPSHOT</version>
+    <packaging>jar</packaging>
+
+    <name>hzau-bbs-service</name>
 
     <dependencies>
         <dependency>
