--- hzau-bbs-web/src/main/java/org/feidian/hzau/bbs/web/HzauBbsWebApplication.java
@@ -1,17 +1,0 @@
-package org.feidian.hzau.bbs.web;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-/**
- * @author stormphoenix
- * @date 2018-09-05
- * @time 下午11:29
- */
-@SpringBootApplication
-public class HzauBbsWebApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(HzauBbsWebApplication.class, args);
-    }
-}
