+++ hzau-bbs-web/src/main/java/org/feidian/hzau/bbs/web/controller/LoginController.java
@@ -1,7 +1,16 @@
 package org.feidian.hzau.bbs.web.controller;
 
+import lombok.Getter;
+import lombok.Setter;
+import org.feidian.hzau.bbs.data.entity.User;
+import org.feidian.hzau.bbs.service.user.UserService;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.servlet.ModelAndView;
+
+import java.util.List;
 
 /**
  * @author stormphoenix
@@ -11,8 +20,17 @@ import org.springframework.web.bind.annotation.RequestMapping;
 @Controller
 public class LoginController extends BaseController {
 
-    @RequestMapping("/")
-    public String index() {
+    @Getter
+    @Setter
+    private List<User> users;
+
+    @Autowired
+    private UserService userService;
+
+    @RequestMapping("/users")
+    public String index(Model model) {
+        users = userService.getAll();
+        model.addAttribute("users", users);
         return "index";
     }
 }
