--- hzau-bbs-web/src/test/java/org/feidian/hzau/bbs/web/HzauBbsWebApplicationTests.java
@@ -1,16 +1,0 @@
-package org.feidian.hzau.bbs.web;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class HzauBbsWebApplicationTests {
-
-    @Test
-    public void contextLoads() {
-    }
-
-}
